check kernel version for raw mode
The downstream Freescale vendor kernel has a patch that allows
determining
if raw NAND flash mode is provided via a debugfs file. This is not
present
in upstream kernels, but the raw access support was added in the 3.19
kernel, so we will check the kernel version if we can't find the file.

Add function to check if kernel support raw_mode and fix function that
calculates the geometry based on known info from the mtd partitionmtd
geometry.

These patches were copied from:
https://github.com/Gateworks/openwrt/commit/379e87df4fb67198fe900ffe2a394e69fbd47801

Signed-off-by: Fabio Berton <fabio.berton@ossystems.com.br>

Signed-off-by: Han Xu <han.xu@nxp.com>