fix the block0 size issue
The Freescale downstream vendor kernel has a patch that exports the bch
flash geometry via a debugfs file. This is not available in mainline
linux
kernels so the fallback method calculates the geometry based on known
info
from the mtd partition. A bug exists in this funcion where it fails to
assume that block0 ECC is the same as the other blocks by default.

Add function to check if kernel support raw_mode and fix function that
calculates the geometry based on known info from the mtd partitionmtd
geometry.

These patches were copied from:
https://github.com/Gateworks/openwrt/commit/379e87df4fb67198fe900ffe2a394e69fbd47801

Signed-off-by: Fabio Berton <fabio.berton@ossystems.com.br>

Signed-off-by: Han Xu <han.xu@nxp.com>