Fix incorrect failing on mtd_erase
In the event where the mtd_erase function also needs to erase the last
block of an mtd device, the erasing goes well but the ofs check causes
the function to fail even though the erasing of the block went well.

This commit ensures that checking whether ofs does not overrun the mtd
size happens as early as possible, thus allowing the mtd_erase function
to return properly when everything went right in erasing the last block
of an mtd device.